reg_app/requirements.txt
streamlit
langchain
openai
pypdf
python-docx
faiss-cpu
sentence-transformers
python-dotenv